+++ include/dmn-blockingqueue.hpp
@@ -191,7 +191,7 @@ public:
    *
    * @return The total number of items that have been passed through the queue.
    */
-  virtual auto waitForEmpty() -> size_t;
+  virtual auto waitForEmpty() -> uint64_t;
 
 protected:
   /**
@@ -223,11 +223,11 @@ private:
   std::condition_variable m_empty_cond{}; // signalled when queue becomes empty
   std::condition_variable
       m_not_empty_cond{}; // signalled on push (multi-pop timed wait)
-  size_t m_push_count{};
-  size_t m_pop_count{};
+  uint64_t m_push_count{};
+  uint64_t m_pop_count{};
   bool m_shutdown{};
 
-  std::atomic<std::size_t> m_popcall_count{};
+  std::atomic<std::uint64_t> m_popcall_count{};
 }; // class Dmn_BlockingQueue
 
 template <typename T> Dmn_BlockingQueue<T>::Dmn_BlockingQueue() {}
@@ -243,7 +243,7 @@ Dmn_BlockingQueue<T>::Dmn_BlockingQueue(std::initializer_list<T> list)
 template <typename T> Dmn_BlockingQueue<T>::~Dmn_BlockingQueue() noexcept try {
   stop();
 
-  size_t popcall_count{};
+  uint64_t popcall_count{};
   while ((popcall_count = m_popcall_count.load(std::memory_order_acquire)) >
          0) {
     m_popcall_count.wait(popcall_count, std::memory_order_acquire);
@@ -330,8 +330,8 @@ template <typename T> void Dmn_BlockingQueue<T>::stop() {
   m_not_empty_cond.notify_all();
 }
 
-template <typename T> auto Dmn_BlockingQueue<T>::waitForEmpty() -> size_t {
-  size_t inbound_count{};
+template <typename T> auto Dmn_BlockingQueue<T>::waitForEmpty() -> uint64_t {
+  uint64_t inbound_count{};
 
   std::unique_lock<std::mutex> lock(m_mutex);
